+++ src/App.tsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import ChickenTable from './components/ChickenTable/ChickenTable';
 import ProductsFilter from './components/Products/ProductsFilter';
 import { ChickenData } from './types/ChickenData';
-import Sidebar from './component/Sidebar/Sidebar';
-import DemoPage from './component/DemoPage';
+import DemoPage from "./components/DemoPage";
+import Sidebar from "./components/Sidebar/Sidebar";
 import './App.css';
 
 const App: React.FC = () => {
